Added recovery, and auto merging capability
Auto Recovery:
==============
The script can now attempt to recover if it fails at some point during
the middle of a synchronization attempt.  Although not fool proof, it
is fairly good about figuring out where it was, and starting from
that point.

Auto Merge
==============
When files have changed in both clearcase and git, the script will try to
merge in the changes automagically using git merge-files.  If this fails,
it will attempt to open up a merge tool (current Perforce's mergetool) to
allow the user to resolve the merge conflicts.  It will then continue on
it's merry way.

Other Fixes
==============
* Better output; stacktraces are dumped when possible.
* Numerous minor issues/typos fixed.
* Prompts added to alert user when bad things could happen, and to
give the user options.

+++ sync.py
@@ -10,26 +10,29 @@ ARGS= {
 }
 
 def main(glob,ignore_whitespace=False):
-    base = abspath(CC_DIR)
-    for i in cfg.getList('include', '.'):
-        for (dirpath, dirnames, filenames) in os.walk(buildPath([CC_DIR, i])):
-            reldir = dirpath[len(base):]
-            for file in filenames:
-                if fnmatch(file, glob):
-                    newFile = buildPath([GIT_DIR, reldir, file])
-                    fromFile = buildPath([dirpath,file])
-                    if (not ignore_whitespace) or doDiff(newFile,fromFile):
-                        if DEBUG: 
-                            print('Copying',fromFile,"to",newFile)
-                        mkdirs(newFile)
-                        shutil.copy(fromFile, newFile)
-                        os.chmod(newFile, stat.S_IWRITE)
+    quest = "This will copy all files matching " + glob + " in the clearcase view to the git repository.\n"
+    quest += "Are you sure you want to do this?"
+
+    if ask(quest,["Yes","No"]) == 'y':
+        base = abspath(CC_DIR)
+        for i in cfg.getList('include', '.'):
+            for (dirpath, dirnames, filenames) in os.walk(buildPath([CC_DIR, i])):
+                reldir = dirpath[len(base):]
+                for file in filenames:
+                    if fnmatch(file, glob):
+                        newFile = buildPath([GIT_DIR, reldir, file])
+                        fromFile = buildPath([dirpath,file])
+                        if (not ignore_whitespace) or doDiff(newFile,fromFile):
+                            if DEBUG: 
+                                print('Copying',fromFile,"to",newFile)
+                            mkdirs(newFile)
+                            shutil.copy(fromFile, newFile)
+                            os.chmod(newFile, stat.S_IWRITE)
 
 def doDiff(file1,file2):
     cmd = ['--ignore-all-space', file1, file2]
     output = popen('diff', cmd, cwd=os.getcwd())
-    if DEBUG:
-        print("Output from diff is", output)
+    debug("Output from diff is: " + output)
     if output:
         return True
     return False
